Refactor method past, models -> controllers
